Trait wae::io::read::AsyncRead

pub trait AsyncRead {
    unsafe fn poll_read(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>, 
        buf: &mut IoSliceMut<'_>
    ) -> Poll<Result<usize>>;

    fn cancel_read(self: Pin<&mut Self>, wait: bool) -> Result<()>;
}

Required methods

unsafe fn poll_read(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
buf: &mut IoSliceMut<'_>
) -> Poll<Result<usize>>

Safety

The given buffer is guaranteed to be valid and stay the same until the function returns Poll::Ready

fn cancel_read(self: Pin<&mut Self>, wait: bool) -> Result<()>

Loading content...