Trait wae::io::write::AsyncWrite

pub trait AsyncWrite {
    unsafe fn poll_write(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>, 
        buf: &IoSlice<'_>
    ) -> Poll<Result<usize>>;

    fn cancel_write(self: Pin<&mut Self>, wait: bool) -> Result<()>;
}

Required methods

unsafe fn poll_write(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
buf: &IoSlice<'_>
) -> Poll<Result<usize>>

Safety

The given buffer is guaranteed to be valid and stay the same until the function returns Poll::Ready

fn cancel_write(self: Pin<&mut Self>, wait: bool) -> Result<()>

Loading content...